fix(start): ensure home callback restores menu
